python_package_requirements.txt
Flask==2.2.3 
pymongo==4.3.3 
gridfs-fuse
web3==5.31.3